+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `@homebridge-plugins/homebridge-matter` will be documente
 - refactor codebase for class-based devices
 - add logging for all `MatterRequest` events
 - fix commands for lock devices
+- update plugin to support `@matter/main@0.16.x`
 
 ## v0.1.6 (2025-10-24)